Time: 1 ms (95.03%) | Memory: 46.9 MB (78.40%) - LeetSync

+++ 3846-minimum-operations-to-make-array-sum-divisible-by-k/minimum-operations-to-make-array-sum-divisible-by-k.java
@@ -6,6 +6,6 @@ class Solution {
 
             s=s+nums[i];
         }
-        return s%k;
+        return (int)s%k;
     }
 }
